Reorganize functions to allow for testing
