+++ src/css/base.scss
@@ -4,7 +4,8 @@
 @import "trip_styles";
 
 body {
-  @include gridDisplay(2fr 8fr 2fr, 1fr 10fr 1fr);
+  @include gridDisplay(2fr 8fr 2fr, 1.5fr 9fr 1.5fr);
+  grid-gap: 10px;
   height: calc(100vh - 20px);
   width: calc(100vw - 20px);
   background: linear-gradient(to bottom right, $black, $darker-purple);
@@ -19,7 +20,8 @@ header {
   grid-column: 1 / -1;
   grid-row: 1 / 2;
   @include headerFooterStyles;
-  box-shadow: 0px 0px 2px 2px whitesmoke;
+  box-shadow: 0px 0px 5px 2px whitesmoke;
+  padding: 10px;
 }
 
 nav {
@@ -41,7 +43,7 @@ main {
   border: $grid-border;
   margin: 3px;
   display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+  grid-template-columns: repeat(auto-fit, minmax(320px, 1.5fr));
   grid-gap: 15px;
   padding: 10px;
   overflow-y: scroll;
