--- src/destinations/destination.js
@@ -1,7 +1,0 @@
-class DestinationRepo {
-  constructor(destination) {
-    
-  }
-}
-
-export default Destination;
